requirements.txt
aioredis==2.0.1
aiosmtplib==2.0.1
alembic==1.9.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.6.2
Flask-Mailing==0.2.1
Flask-Migrate==4.0.0
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gtfs-realtime-bindings==0.0.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
protobuf==4.21.12
pydantic==1.10.4
pyscopg2==66.0.2
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
requests-cache==0.9.7
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
tomli==2.0.1
typing_extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
